+++ src/sources/source_datastream.jl
@@ -22,9 +22,9 @@ function query{S<:DataStreams.Data.Source}(source::S)
         if schema.types[i] <: WeakRefString
             col_type = String
         elseif schema.types[i] <: Nullable && schema.types[i].parameters[1] <: WeakRefString
-            col_type = NAable{String}
+            col_type = DataValue{String}
         elseif schema.types[i] <: Nullable
-            col_type = NAable{schema.types[i].parameters[1]}
+            col_type = DataValue{schema.types[i].parameters[1]}
         else
             col_type = schema.types[i]
         end
@@ -64,9 +64,9 @@ end
 function _convertion_helper_for_datastreams(source, row, col, T)
     v = Data.streamfrom(source, Data.Field, Nullable{T}, row, col)
     if isnull(v)
-        return NAable{String}()
+        return DataValue{String}()
     else
-        return NAable{String}(String(get(v)))
+        return DataValue{String}(String(get(v)))
     end
 end
 
@@ -75,10 +75,10 @@ end
     for i in 1:length(TC.types)
         if TC.types[i] <: String
             get_expression = :(Data.streamfrom(source, Data.Field, WeakRefString, row, $i))
-        elseif TC.types[i] <: NAable && TSC.types[i].parameters[1] <: WeakRefString
+        elseif TC.types[i] <: DataValue && TSC.types[i].parameters[1] <: WeakRefString
             get_expression = :(_convertion_helper_for_datastreams(source, row, $i, TSC.types[$i].parameters[1]))
-        elseif TC.types[i] <: NAable
-            get_expression = :(NAable(Data.streamfrom(source, Data.Field, Nullable{$(TC.types[i].parameters[1])}, row, $i)))
+        elseif TC.types[i] <: DataValue
+            get_expression = :(DataValue(Data.streamfrom(source, Data.Field, Nullable{$(TC.types[i].parameters[1])}, row, $i)))
         else
             get_expression = :(Data.streamfrom(source, Data.Field, $(TC.types[i]), row, $i))
         end