+++ src/sources/source_typedtable.jl
@@ -13,7 +13,7 @@ function query(df::TypedTables.Table)
     df_columns_tuple_type = Expr(:curly, :Tuple)
     for i in 1:length(df.data)
         etype = eltype(df.data[i])
-        push!(col_expressions, Expr(:(::), names(df)[i], etype <: Nullable ? NAable{etype.parameters[1]} : etype))
+        push!(col_expressions, Expr(:(::), names(df)[i], etype <: Nullable ? DataValue{etype.parameters[1]} : etype))
         push!(df_columns_tuple_type.args, typeof(df.data[i]))
     end
     t_expr = NamedTuples.make_tuple(col_expressions)
@@ -45,7 +45,7 @@ end
 @generated function next{T,TS}(iter::EnumerableTypedTable{T,TS}, state)
     constructor_call = Expr(:call, :($T))
     for (i,t) in enumerate(T.parameters)
-        push!(constructor_call.args, t<:NAable ? :(isnull(columns[$i][i]) ? NAable{$(t.parameters[1])}() : NAable{$(t.parameters[1])}(get(columns[$i][i]))) : :(columns[$i][i]))
+        push!(constructor_call.args, t<:DataValue ? :(isnull(columns[$i][i]) ? DataValue{$(t.parameters[1])}() : DataValue{$(t.parameters[1])}(get(columns[$i][i]))) : :(columns[$i][i]))
     end
 
     quote